+++ src/main.rs
@@ -11,6 +11,7 @@ mod theme;
 mod ui_add_proc;
 mod ui_keymap;
 mod ui_procs;
+mod ui_remove_proc;
 mod ui_term;
 
 use std::path::Path;